+++ pkg/armrpc/frontend/defaultcontroller/getoperationresult.go
