+++ pkg/armrpc/frontend/defaultcontroller/testdata/operationstatus_datamodel.json
